+++ idea/testData/injection/InjectionOfCustomParameterJavaWithAnnotation.kt
@@ -1,0 +1,3 @@
+import some.Test
+
+fun bar() { Test.foo("<caret>some") }