+++ ChangeLog.md
@@ -19,7 +19,7 @@ New features:
 - [KT-2428](https://youtrack.jetbrains.com/issue/KT-11574) Support basic use-cases of language injection for expressions marked with @Language annotation
 - [KT-11472](https://youtrack.jetbrains.com/issue/KT-11472) Add comment or @Language annotation after "Inject language or reference" intention automatically
 - Apply injection for the literals in property initializer through property usages
-- Enable injection in Kotlin by annotation function or method parameter with @Language annotation
+- Enable injection from Java or Kotlin function declaration by annotating parameter with @Language annotation
 - [KT-11807](https://youtrack.jetbrains.com/issue/KT-11807) Use function body template when generating overriding functions with default body
 
 Issues fixed: